Pretransformations to use destination passing style, before upstream bufferization. (#7582)
The upstream ComprehensiveBufferization pass prefers a more
destination-passing style of code, i.e. it has a way to know before
hand what the output buffer to use is for in-place
operations. Specifically, the use of init_tensor to conjure a tensor
(even though it might be fully over-written before its values are
used) results in an allocation. It is easier to do a
pre-transformation that removes these init_tensor operations and
replaces them with a flow.dispatch.tensor.load of the output where
the final result goes into.
The current IREE bufferization handles this. This patch reuses the
relevant parts of IREE bufferization to effect this transformation.
