+++ iree/compiler/Codegen/Common/test/BUILD
@@ -21,6 +21,7 @@ iree_lit_test_suite(
         [
             "affinemin_canonicalization.mlir",
             "canonicalize_interface_load_store.mlir",
+            "convert_to_destination_passing_style.mlir",
             "dead_alloc.mlir",
             "f32Tof16.mlir",
             "flatten_memref_subspan.mlir",
