+++ iree/compiler/Codegen/Common/test/CMakeLists.txt
@@ -16,6 +16,7 @@ iree_lit_test_suite(
   SRCS
     "affinemin_canonicalization.mlir"
     "canonicalize_interface_load_store.mlir"
+    "convert_to_destination_passing_style.mlir"
     "dead_alloc.mlir"
     "f32Tof16.mlir"
     "flatten_memref_subspan.mlir"
